feature/results
adding context,
adding results data
adding results display logic
