+++ src/app/components/question-form/index.ts
@@ -1,1 +1,1 @@
-export * from './QuestionForm'
+export {default} from './QuestionForm'
