+++ src/app/components/result/index.ts
@@ -1,0 +1,1 @@
+export {default} from './Result'
