+++ src/app/test/Test.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 
-import { Box } from "@mui/material";
 import { BoxTestStyle } from "./TestStyle";
-import QuestionForm from "../components/question-form/QuestionForm";
+import QuestionForm from "../components/question-form";
 
 const Test = () => {
   return (
